feat: add Cloud Build configuration for building and deploying the Langfuse app
This commit adds a Cloud Build configuration file that specifies the steps for building and deploying the Langfuse app. The steps include building the Docker image, pushing it to a container registry, and deploying it to Google Cloud Run. The image that is built and deployed is "europe-west1-docker.pkg.dev/langfuse-test/langfuse/langfuse:latest".