WIP on multi-prompt playground
The Context and Cache of the playground require refactoring such that instead of creating a full playground provider per open prompt and a separate sessionStorage cache key per prompt they are handled in one context managing an array of prompts which is then also serialized as a JSON array into the cache. Furthermore the different actions re-used from the single page need to be adapted to the new context model once that has been reworked. Also there is functionality missing to duplicate prompts, load prompts from the backend and to select versions as well as hint at unsaved changes. CSS around the sizing of the prompt components depending on how many are present in the list is also missing.